requirements.txt
openai>=1.0.0
pdfplumber>=0.9.0
tiktoken>=0.5.0
faiss-cpu>=1.7.4
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
streamlit>=1.28.0
